Done: pushed version of miniRT

+++ include/point.h
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   point.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: habouiba <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: atarchou <atarchou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 17:58:10 by habouiba          #+#    #+#             */
-/*   Updated: 2022/11/15 18:00:22 by habouiba         ###   ########.fr       */
+/*   Updated: 2023/01/07 00:08:36 by atarchou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef POINT_H
-#define POINT_H
+# define POINT_H
 
-#include "types.h"
+# include "types.h"
 
-t_point *point_create(float x, float y, float z);
+t_point	*point_create(double x, double y, double z);
 #endif
