+++ src/core/matrix/delete.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   delete.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: habouiba <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: atarchou <atarchou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/18 15:11:47 by habouiba          #+#    #+#             */
-/*   Updated: 2022/11/22 10:39:24 by habouiba         ###   ########.fr       */
+/*   Updated: 2023/01/07 00:08:36 by atarchou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "matrix.h"
 
-void matrix_free_n(void *_m, int n)
+void	matrix_free_n(void *_m, int n)
 {
-	float **m;
-	int     i;
+	double	**m;
+	int		i;
 
 	m = _m;
 	i = 0;
@@ -27,17 +27,17 @@ void matrix_free_n(void *_m, int n)
 	free(m);
 }
 
-void matrix_free_4(void *_m)
+void	matrix_free_4(void *_m)
 {
 	matrix_free_n(_m, 4);
 }
 
-void matrix_free_3(void *_m)
+void	matrix_free_3(void *_m)
 {
 	matrix_free_n(_m, 3);
 }
 
-void matrix_free_2(void *_m)
+void	matrix_free_2(void *_m)
 {
 	matrix_free_n(_m, 2);
 }
