+++ src/core/matrix/scale.c
@@ -6,15 +6,15 @@
 /*   By: atarchou <atarchou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 11:01:07 by habouiba          #+#    #+#             */
-/*   Updated: 2022/12/02 21:17:08 by atarchou         ###   ########.fr       */
+/*   Updated: 2023/01/07 00:08:36 by atarchou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "matrix.h"
 
-float **matrix_scale_4(t_tuple *t, void (*f)(void *))
+double	**matrix_scale_4(t_tuple *t, void (*f)(void *))
 {
-	float **scaling_m;
+	double	**scaling_m;
 
 	scaling_m = matrix_identity_4();
 	scaling_m[0][0] = t->x;
