+++ src/core/matrix/translate.c
@@ -6,16 +6,16 @@
 /*   By: atarchou <atarchou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/20 18:09:41 by habouiba          #+#    #+#             */
-/*   Updated: 2022/12/02 21:16:37 by atarchou         ###   ########.fr       */
+/*   Updated: 2023/01/07 00:08:36 by atarchou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "matrix.h"
 #include <stdio.h>
 
-float **matrix_translate_creat_4(t_tuple *t, void (*f)(void *))
+double	**matrix_translate_creat_4(t_tuple *t, void (*f)(void *))
 {
-	float **translate;
+	double	**translate;
 
 	translate = matrix_identity_4();
 	translate[0][3] = t->x;
