+++ src/core/point/point.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   point.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: habouiba <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: atarchou <atarchou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 17:59:15 by habouiba          #+#    #+#             */
-/*   Updated: 2022/11/15 18:01:01 by habouiba         ###   ########.fr       */
+/*   Updated: 2023/01/07 00:08:36 by atarchou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "point.h"
 #include "tuple.h"
 
-t_point *point_create(float x, float y, float z)
+t_point	*point_create(double x, double y, double z)
 {
-	t_point *p;
+	t_point	*p;
 
 	p = tuple_create(x, y, z, 1);
 	return (p);
