+++ src/core/ray/at.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   at.c                                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: habouiba <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: atarchou <atarchou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 12:15:12 by habouiba          #+#    #+#             */
-/*   Updated: 2022/11/21 12:33:03 by habouiba         ###   ########.fr       */
+/*   Updated: 2023/01/07 00:08:36 by atarchou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 #include "ray.h"
 #include "tuple.h"
 
-t_point *ray_at(t_ray *ray, float t)
+t_point	*ray_at(t_ray *ray, double t)
 {
-	t_point *scaled;
-	t_point *p;
+	t_point	*scaled;
+	t_point	*p;
 
 	scaled = tuple_scale(ray->dir, t);
 	p = tuple_add(ray->origin, scaled);
