+++ src/renderer/hit/hit.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   hit.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: habouiba <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: atarchou <atarchou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 17:09:10 by habouiba          #+#    #+#             */
-/*   Updated: 2022/11/22 13:43:32 by habouiba         ###   ########.fr       */
+/*   Updated: 2023/01/07 00:08:36 by atarchou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,14 +17,14 @@
  * take a list of hit and returns the closest one
  */
 
-float which_positive_and_low(float a, float b)
+double	which_positive_and_low(double a, double b)
 {
 	if (a >= 0.0f && a < b)
 		return (a);
 	return (b);
 }
 
-t_hit *which_is_close_to_cam(t_hit *low, t_hit *curr)
+t_hit	*which_is_close_to_cam(t_hit *low, t_hit *curr)
 {
 	if (curr->t1 < 0.0f && curr->t2 < 0.0f)
 		return (low);
@@ -42,14 +42,13 @@ t_hit *which_is_close_to_cam(t_hit *low, t_hit *curr)
 		if (which_positive_and_low(low->t1, low->t2) > curr->t2)
 			return (curr);
 	}
-	// one if negative or both positive
 	return (low);
 }
 
-t_hit *hit(t_list *lst)
+t_hit	*hit(t_list *lst)
 {
-	t_hit *low;
-	t_hit *curr;
+	t_hit	*low;
+	t_hit	*curr;
 
 	low = 0;
 	while (lst)
